fix: refactor rating to use postgres, fix bugs. TODO add event service

+++ rating-service/src/main/kotlin/ru/aasmc/ratingservice/appevents/listener/FilmLikeEventListener.kt
@@ -44,7 +44,7 @@ class FilmLikeEventListener(
                 val filmLike = FilmLike(
                         id = id,
                         timestamp = event.dto.timestamp,
-                        mark = event.dto.mark
+                        mark = event.dto.mark.toDouble()
                 )
                 filmLikeService.saveFilmLike(filmLike)
             }
